+++ src/Service/PeopleLinkService.php
@@ -26,82 +26,6 @@ class PeopleLinkService
     $this->request = $requestStack->getCurrentRequest();
   }
 
-  public function discoveryLink(People $company, People $people, $linkType): PeopleLink
-  {
-    $peopleLink = $this->manager->getRepository(PeopleLink::class)->findOneBy([
-      'company' => $company,
-      'people' => $people,
-      'link_type' => $linkType
-    ]);
-
-    if (!$peopleLink)
-      $peopleLink = $this->addLink($company, $people, $linkType);
-
-    return $peopleLink;
-  }
-
-  public function addLink(People $company, People $people, $link_type): PeopleLink
-  {
-    $peopleLink = $this->manager->getRepository(PeopleLink::class)->findOneBy([
-      'company' => $company,
-      'people' => $people,
-      'link_type' => $link_type
-    ]);
-
-    if (!$peopleLink)
-      $peopleLink = new PeopleLink();
-
-    $peopleLink->setCompany($company);
-    $peopleLink->setPeople($people);
-    $peopleLink->setLinkType($link_type);
-
-    $this->manager->persist($peopleLink);
-    $this->manager->flush();
-
-    return $peopleLink;
-  }
-
-  public function securityFilter(QueryBuilder $queryBuilder, $resourceClass = null, $applyTo = null, $rootAlias = null): void
-  {
-    $this->checkLink($queryBuilder, $rootAlias);
-  }
-
-  public function checkLink(QueryBuilder $queryBuilder, $rootAlias): void
-  {
-    $link = $this->request->query->get('link', null);
-    $company = $this->request->query->get('company', null);
-    $link_type = $this->request->query->get('link_type', null);
-
-    if ($link_type) {
-      $queryBuilder->join(sprintf('%s.' . ($link ? 'company' : 'link'), $rootAlias), 'PeopleLink');
-      $queryBuilder->andWhere('PeopleLink.link_type IN(:link_type)');
-      $queryBuilder->setParameter('link_type', $link_type);
-    }
-
-    if ($company || $link) {
-      $queryBuilder->andWhere('PeopleLink.' . ($link ? 'people' : 'company') . ' IN(:people)');
-      $queryBuilder->setParameter('people', preg_replace("/[^0-9]/", "", ($link ?: $company)));
-    }
-  }
-
-  public function getMyCompanies(): array
-  {
-    $token = $this->security->getToken();
-    if (!$token) return [];
-
-    $currentUser = $token->getUser();
-    $companies = [];
-    if (!$currentUser) return [];
-
-    if (!$currentUser->getPeople()->getLink()->isEmpty()) {
-      foreach ($currentUser->getPeople()->getLink() as $company) {
-        if ($company->getLinkType() == 'employee')
-          $companies[] = $company->getCompany();
-      }
-    }
-
-    return $companies;
-  }
 
 
   public function notifyClient(PeopleLink $peopleLink): ?PeopleLink
